# docs/swagger.yaml
definitions:
  request.Login:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  request.User:
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        maxLength: 20
        minLength: 6
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      phone:
        maxLength: 11
        minLength: 11
        type: string
      username:
        maxLength: 20
        minLength: 6
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg: {}
    type: object
info:
  contact: {}
paths:
  /user:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List user by id
      tags:
      - User
    post:
      parameters:
      - description: User Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Post user by id
      tags:
      - User
  /user/{id}:
    delete:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete user by id
      tags:
      - User
    get:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get user by id
      tags:
      - User
    patch:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      - description: User Info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Patch user by id
      tags:
      - User
  /user/login:
    post:
      parameters:
      - description: Login Info
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user with username and password
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    description: JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"